# swagger.yaml
openapi: 3.0.3
info:
  title: 'AXZY CHECK API'
  description: "Documentación técnica completa de la API de AXZY CHECK. Cubre gestión\noperativa, administrativa y analítica."
  version: 1.3.2
servers:
  -
    url: 'http://localhost:4444/api/v1'
    description: 'Local server'
...
paths:
...
  /incidents:
    get:
      tags:
        - incidents
      summary: 'Listar incidencias'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResult'
    post:
      tags:
        - incidents
      summary: 'Crear reporte de incidencia'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentCreate'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResult'
        '400':
          description: 'Datos inválidos'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResultError'
...
  '/incidents/{id}':
    delete:
      tags:
        - incidents
      summary: 'Eliminar incidencia'
      parameters:
        -
          in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResultListRound'
  /maintenance:
    get:
      tags:
        - maintenance
      summary: 'Listar solicitudes'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResultListMaintenance'
    post:
      tags:
        - maintenance
      summary: 'Crear solicitud'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceCreate'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResultDatatableRound'
        '400':
          description: 'Datos inválidos'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResultError'
...
  '/maintenance/{id}':
    delete:
      tags:
        - maintenance
      summary: 'Eliminar mantenimiento'
      parameters:
        -
          in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResultDatatableIncident'
        '400':
          description: 'Petición inválida o restricción de llave foránea'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResultError'
        '404':
          description: 'No encontrado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResultError'
        '500':
          description: 'Error interno del servidor'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResultError'
...
  /kardex:
    get:
      tags:
        - kardex
      summary: 'Listar bitácora'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResultDatatableSchedule'
        '400':
          description: 'Petición inválida'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResultError'
        '404':
          description: 'No encontrado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResultError'
        '500':
          description: 'Error interno del servidor'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResultError'
    post:
      tags:
        - kardex
      summary: 'Crear registro'
      responses:
        '201':
          description: Creado
...
  '/kardex/{id}':
    get:
      tags:
        - kardex
      summary: 'Detalle de registro Kardex'
      parameters:
        -
          in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResultKardex'
        '404':
          description: 'No encontrado'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResultError'
    patch:
      tags:
        - kardex
      summary: 'Actualizar registro Kardex'
      parameters:
        -
          in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResultKardex'
    delete:
      tags:
        - kardex
      summary: 'Eliminar registro Kardex'
      parameters:
        -
          in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TResult'
...
components:
  schemas:
...
    ClientBasic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
          nullable: true
        rfc:
          type: string
          nullable: true
        contactName:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        active:
          type: boolean
        softDelete:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
...
    IncidentCreate:
      type: object
      required:
        - title
        - locationId
      properties:
        title:
          type: string
        description:
          type: string
        locationId:
          type: string
          format: uuid
        media:
          type: array
          items:
            type: string
    MaintenanceCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        description:
          type: string
        media:
          type: array
          items:
            type: string
...
    Round:
      type: object
      properties:
        id:
          type: string
          format: uuid
        guardId:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
          nullable: true
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/RoundStatus'
        recurringConfigurationId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
        guard:
          $ref: '#/components/schemas/UserBasic'
        client:
          $ref: '#/components/schemas/ClientBasic'
        recurringConfiguration:
          $ref: '#/components/schemas/RecurringConfigurationBasic'
...
    Incident:
      type: object
      properties:
        id:
          type: string
          format: uuid
        guardId:
          type: string
          format: uuid
        title:
          type: string
        categoryId:
          type: string
          format: uuid
          nullable: true
        typeId:
          type: string
          format: uuid
          nullable: true
        description:
          type: string
          nullable: true
        media:
          type: object
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        resolvedById:
          type: string
          format: uuid
          nullable: true
        status:
          $ref: '#/components/schemas/IncidentStatus'
        clientId:
          type: string
          format: uuid
          nullable: true
        client:
          $ref: '#/components/schemas/ClientBasic'
        category:
          $ref: '#/components/schemas/IncidentCategoryBasic'
        type:
          $ref: '#/components/schemas/IncidentTypeBasic'
        guard:
          $ref: '#/components/schemas/UserBasic'
        resolvedBy:
          $ref: '#/components/schemas/UserBasic'
    Maintenance:
      type: object
      properties:
        id:
          type: string
          format: uuid
        guardId:
          type: string
          format: uuid
        title:
          type: string
        categoryId:
          type: string
          format: uuid
          nullable: true
        typeId:
          type: string
          format: uuid
          nullable: true
        category:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        media:
          type: object
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        status:
          $ref: '#/components/schemas/MaintenanceStatus'
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        resolvedById:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
        guard:
          $ref: '#/components/schemas/UserBasic'
        resolvedBy:
          $ref: '#/components/schemas/UserBasic'
        categoryRel:
          $ref: '#/components/schemas/IncidentCategoryBasic'
        type:
          $ref: '#/components/schemas/IncidentTypeBasic'
        clientId:
          type: string
          format: uuid
          nullable: true
        client:
          $ref: '#/components/schemas/ClientBasic'
    Kardex:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        locationId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
        media:
          type: object
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        assignmentId:
          type: string
          format: uuid
          nullable: true
        scanType:
          $ref: '#/components/schemas/ScanType'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
        assignment:
          $ref: '#/components/schemas/AssignmentBasic'
        location:
          $ref: '#/components/schemas/LocationBasic'
        user:
          $ref: '#/components/schemas/UserBasic'
    TResult:
      type: object
      description: 'Estructura de respuesta estándar AXZY'
      properties:
        success:
          type: boolean
          description: 'Indica si la operación fue exitosa'
          example: true
        data:
          type: object
          description: 'Datos de la respuesta'
          example:
            id: uuid
            name: Ejemplo
        messages:
          type: array
          description: 'Lista de mensajes informativos o de éxito'
          items:
            type: string
          example:
            - 'Operación realizada con éxito'
...
    TResultError:
      type: object
      description: 'Estructura de respuesta de error estándar'
      properties:
        success:
          type: boolean
          example: false
        data:
          type: object
          nullable: true
          example: null
        messages:
          type: array
          description: 'Detalles del error o validación'
          items:
            type: string
          example:
            - 'Credenciales inválidas'
            - 'El campo nombre es obligatorio'
        stack:
          type: string
          description: 'Traza del error (solo en desarrollo)'
          example: 'Error: Stack trace sample...'
...
    TResultDatatableRound:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            rows:
              type: array
              items:
                $ref: '#/components/schemas/Round'
            total:
              type: integer
              example: 100
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
        messages:
          type: array
          items:
            type: string
    TResultListRound:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Round'
        messages:
          type: array
          items:
            type: string
    TResultDatatableIncident:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            rows:
              type: array
              items:
                $ref: '#/components/schemas/Incident'
            total:
              type: integer
              example: 100
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
        messages:
          type: array
          items:
            type: string
...
    TResultListMaintenance:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Maintenance'
        messages:
          type: array
          items:
            type: string
    Schedule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserBasic'
        shiftPlans:
          type: array
          items:
            $ref: '#/components/schemas/ShiftPlanBasic'
        shiftHandovers:
          type: array
          items:
            $ref: '#/components/schemas/ShiftHandoverBasic'
        uniformChecks:
          type: array
          items:
            $ref: '#/components/schemas/UniformCheckBasic'
    TResultDatatableSchedule:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            rows:
              type: array
              items:
                $ref: '#/components/schemas/Schedule'
            total:
              type: integer
              example: 100
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
        messages:
          type: array
          items:
            type: string
...
    ScanType:
      type: string
      enum:
        - ASSIGNMENT
        - RECURRING
        - FREE
    IncidentStatus:
      type: string
      enum:
        - PENDING
        - ATTENDED
    RoundStatus:
      type: string
      enum:
        - IN_PROGRESS
        - COMPLETED
    MaintenanceStatus:
      type: string
      enum:
        - PENDING
        - ATTENDED
...
    UserBasic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        lastName:
          type: string
          nullable: true
        username:
          type: string
        active:
          type: boolean
        softDelete:
          type: boolean
        shiftStart:
          type: string
          nullable: true
        shiftEnd:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
        fcmToken:
          type: string
          nullable: true
        isLoggedIn:
          type: boolean
        scheduleId:
          type: string
          format: uuid
          nullable: true
        clientId:
          type: string
          format: uuid
          nullable: true
        roleId:
          type: string
          format: uuid
...
    TResultKardex:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Kardex'
        messages:
          type: array
          items:
            type: string
...
    IncidentCategoryBasic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        value:
          type: string
        type:
          type: string
        color:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
...
    IncidentTypeBasic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        name:
          type: string
        value:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
...
    RecurringConfigurationBasic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        clientId:
          type: string
          format: uuid
          nullable: true
        active:
          type: boolean
        softDelete:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
...
    ShiftPlanBasic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        scheduleId:
          type: string
          format: uuid
        requireHandover:
          type: boolean
        requireUniform:
          type: boolean
        toleranceMinutes:
          type: integer
        daysOfWeek:
          type: array
          items:
            type: integer
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
          nullable: true
...
    ShiftHandoverBasic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        scheduleId:
          type: string
          format: uuid
        shiftDate:
          type: string
          format: date-time
        credentialsCount:
          type: integer
          nullable: true
        tarjetonesCount:
          type: integer
          nullable: true
        novedades:
          type: string
          nullable: true
...
    UniformCheckBasic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        guardId:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
          nullable: true
        scheduleId:
          type: string
          format: uuid
          nullable: true
        shiftDate:
          type: string
          format: date-time
        evaluatedById:
          type: string
          format: uuid
...